requirements.txt
timm==0.6.5
torcheval==0.0.5
datasets==2.21.0
gcsfs==2023.6.0
fsspec==2023.6.0
peft==0.15
sentence-transformers==4.0
transformers==4.52
accelerate==1.6.0
loralib
peft
prodigyopt
dadaptation
dog-optimizer
spacy
tqdm
tensorboard
progress
